bump(asp3): bump up TOPPERS/ASP3 to 3.7.0.
Based on
https://www.toppers.jp/download.cgi/asp3_nucleo_f401re_gcc-20230401.tar.gz

But this generates errors and need fixing.
```
tecsgen  version 1.8.0
../../asp3/../drivers/drivers.cdl:16:31: error: S1042 call port 'cSIOCBR' not initialized in cell 'SIOPortPybricksUSB1'
../../asp3/../drivers/drivers.cdl:19:37: error: S1042 call port 'cSIOCBR' not initialized in cell 'SIOPortPybricksBluetooth1'
../../asp3/../drivers/drivers.cdl:16:31: warning: W1007 ::SIOPortPybricksUSB1 : non-active cell has no entry join and no factory
../../asp3/../drivers/drivers.cdl:19:37: warning: W1007 ::SIOPortPybricksBluetooth1 : non-active cell has no entry join and no factory
2 errors  2 warnings
```

+++ asp3/test/testexec.rb
@@ -4,7 +4,7 @@
 #  TOPPERS Software
 #      Toyohashi Open Platform for Embedded Real-Time Systems
 # 
-#  Copyright (C) 2016-2019 by Embedded and Real-Time Systems Laboratory
+#  Copyright (C) 2016-2023 by Embedded and Real-Time Systems Laboratory
 #              Graduate School of Information Science, Nagoya Univ., JAPAN
 # 
 #  上記著作権者は，以下の(1)〜(4)の条件を満たす場合に限り，本ソフトウェ
@@ -36,8 +36,38 @@
 #  アの利用により直接的または間接的に生じたいかなる損害に関しても，そ
 #  の責任を負わない．
 # 
-#  $Id: testexec.rb 1275 2019-10-03 16:01:48Z ertl-hiro $
+#  $Id: testexec.rb 1790 2023-01-18 06:07:25Z ertl-hiro $
 # 
+
+# 【実行方法】
+#	testexec <処理内容> <処理対象>
+#
+#	処理内容：
+#		デフォルト		buildとexec
+#		build			ビルドのみ
+#		exec			実行のみ
+#		clean			クリーン処理
+#
+#	処理対象：
+#		デフォルト		kernelとall
+#		kernel			ディレクトリが作られているすべてのカーネル
+#		kernel<数字>	指定したビルドオプションのカーネル
+#		all				ディレクトリが作られているすべてのテストプログラム
+#		<テスト名>		指定したテストプログラム
+#
+# 【ターゲット毎のビルドオプション】
+#	ターゲット毎のビルドオプションを，TARGET_OPTIONSに作成する．異なる
+#	テスト用のビルドオプションを，各行に記述する．
+#
+#	各行（最初の行が0）に記述するビルドオプション：
+#		0		機能テストプログラム
+#		1		性能評価プログラム
+#		2		タイマドライバシミュレータを用いたテストプログラム
+#		3		FPUを使用するテストプログラム（ARM向け）
+#
+# 【ターゲット毎の実行方法】
+#	ターゲット上でテストプログラムを実行するための記述を，TARGET_RUNに
+#	作成する．
 
 Encoding.default_external = 'utf-8'
 require "pp"
@@ -63,15 +93,17 @@ TEST_SPEC = {
   "flg1"     => { SRC: "test_flg1" },
   "hrt1"     => { SRC: "test_hrt1" },
   "int1"     => { SRC: "test_int1" },
-  "mutex1"   => { SRC: "test_mutex1", CDL: "test_pf_bitmutex" },
-  "mutex2"   => { SRC: "test_mutex2", CDL: "test_pf_bitmutex" },
-  "mutex3"   => { SRC: "test_mutex3", CDL: "test_pf_bitmutex" },
-  "mutex4"   => { SRC: "test_mutex4", CDL: "test_pf_bitmutex" },
-  "mutex5"   => { SRC: "test_mutex5", CDL: "test_pf_bitmutex" },
-  "mutex6"   => { SRC: "test_mutex6", CDL: "test_pf_bitmutex" },
-  "mutex7"   => { SRC: "test_mutex7", CDL: "test_pf_bitmutex" },
-  "mutex8"   => { SRC: "test_mutex8", CDL: "test_pf_bitmutex" },
+  "mpf1"     => { SRC: "test_mpf1" },
+  "mutex1"   => { SRC: "test_mutex1", CDL: "test_pf_bitkernel" },
+  "mutex2"   => { SRC: "test_mutex2", CDL: "test_pf_bitkernel" },
+  "mutex3"   => { SRC: "test_mutex3", CDL: "test_pf_bitkernel" },
+  "mutex4"   => { SRC: "test_mutex4", CDL: "test_pf_bitkernel" },
+  "mutex5"   => { SRC: "test_mutex5", CDL: "test_pf_bitkernel" },
+  "mutex6"   => { SRC: "test_mutex6", CDL: "test_pf_bitkernel" },
+  "mutex7"   => { SRC: "test_mutex7", CDL: "test_pf_bitkernel" },
+  "mutex8"   => { SRC: "test_mutex8", CDL: "test_pf_bitkernel" },
   "notify1"  => { SRC: "test_notify1" },
+  "pdq1"     => { SRC: "test_pdq1" },
   "raster1"  => { SRC: "test_raster1", CDL: "test_pf_bitkernel" },
   "raster2"  => { SRC: "test_raster2" },
   "sem1"     => { SRC: "test_sem1" },
@@ -89,7 +121,7 @@ TEST_SPEC = {
   # オーバランハンドラ機能拡張パッケージの機能テストプログラム
   "ovrhdr1"  => { SRC: "test_ovrhdr1" },
   "ovrhdr2"  => { SRC: "test_ovrhdr2" },
-  "ovrhdr3"  => { TARGET: 1, SRC: "simt_ovrhdr3", DEFS: "-DHRT_CONFIG1" },
+  "ovrhdr3"  => { TARGET: 2, SRC: "simt_ovrhdr3", DEFS: "-DHRT_CONFIG1" },
   "ovrhdr4"  => { SRC: "test_ovrhdr4" },
 
   # 制約タスク拡張パッケージの機能テストプログラム
@@ -99,69 +131,95 @@ TEST_SPEC = {
   # サブ優先度機能拡張パッケージの機能テストプログラム
   "subprio1" => { SRC: "test_subprio1" },
   "subprio2" => { SRC: "test_subprio2" },
+  "subprio3" => { SRC: "test_subprio3" },
+
+  # 優先度継承拡張パッケージの機能テストプログラム
+  "inherit1" => { SRC: "test_inherit1", CDL: "test_pf_bitkernel" },
+  "inherit2" => { SRC: "test_inherit2", CDL: "test_pf_bitkernel" },
+  "inherit3" => { SRC: "test_inherit3", CDL: "test_pf_bitkernel" },
+  "inherit4" => { SRC: "test_inherit4", CDL: "test_pf_bitkernel" },
+  "inherit5" => { SRC: "test_inherit5", CDL: "test_pf_bitkernel" },
+  "inherit6" => { SRC: "test_inherit6", CDL: "test_pf_bitkernel" },
+  "inherit7" => { SRC: "test_inherit7", CDL: "test_pf_bitkernel" },
 
   # システム時刻管理機能テストプログラム
-  "systim1" => { TARGET: 1, SRC: "simt_systim1",
+  "systim1" => { TARGET: 2, SRC: "simt_systim1",
 								DEFS: "-DHRT_CONFIG1 -DHOOK_HRT_EVENT" },
-  "systim2" => { TARGET: 1, SRC: "simt_systim2",
+  "systim2" => { TARGET: 2, SRC: "simt_systim2",
 								DEFS: "-DHRT_CONFIG1 -DHOOK_HRT_EVENT" },
-  "systim3" => { TARGET: 1, SRC: "simt_systim3",
+  "systim3" => { TARGET: 2, SRC: "simt_systim3",
 								DEFS: "-DHRT_CONFIG1 -DHOOK_HRT_EVENT" },
-  "systim4" => { TARGET: 1, SRC: "simt_systim4",
+  "systim4" => { TARGET: 2, SRC: "simt_systim4",
 								DEFS: "-DHRT_CONFIG2 -DHOOK_HRT_EVENT" },
-  "systim1_64hrt" => { TARGET: 1, SRC: "simt_systim1_64hrt",
+  "systim1_64hrt" => { TARGET: 2, SRC: "simt_systim1_64hrt",
 				CFG: "simt_systim1", DEFS: "-DHRT_CONFIG3 -DHOOK_HRT_EVENT" },
-  "systim2_64hrt" => { TARGET: 1, SRC: "simt_systim2_64hrt",
+  "systim2_64hrt" => { TARGET: 2, SRC: "simt_systim2_64hrt",
 				CFG: "simt_systim2", DEFS: "-DHRT_CONFIG3 -DHOOK_HRT_EVENT" },
-  "systim3_64hrt" => { TARGET: 1, SRC: "simt_systim3_64hrt",
+  "systim3_64hrt" => { TARGET: 2, SRC: "simt_systim3_64hrt",
 				CFG: "simt_systim3", DEFS: "-DHRT_CONFIG3 -DHOOK_HRT_EVENT" },
 
   # ドリフト調整機能拡張パッケージのシステム時刻管理機能テストプログラム
-  "drift1"        => { TARGET: 1, SRC: "simt_drift1",
+  "drift1"        => { TARGET: 2, SRC: "simt_drift1",
 								DEFS: "-DHRT_CONFIG1 -DHOOK_HRT_EVENT" },
-  "drift1_64hrt"  => { TARGET: 1, SRC: "simt_drift1_64hrt",
+  "drift1_64hrt"  => { TARGET: 2, SRC: "simt_drift1_64hrt",
 				CFG: "simt_drift1", DEFS: "-DHRT_CONFIG3 -DHOOK_HRT_EVENT" },
-  "drift1_64ops"  => { TARGET: 1, SRC: "simt_drift1",
+  "drift1_64ops"  => { TARGET: 2, SRC: "simt_drift1",
 				DEFS: "-DHRT_CONFIG1 -DHOOK_HRT_EVENT -DUSE_64BIT_OPS" },
-  "systim1_64ops" => { TARGET: 1, SRC: "simt_systim1",
+  "systim1_64ops" => { TARGET: 2, SRC: "simt_systim1",
 				DEFS: "-DHRT_CONFIG1 -DHOOK_HRT_EVENT -DUSE_64BIT_OPS" },
-  "systim2_64ops" => { TARGET: 1, SRC: "simt_systim2",
+  "systim2_64ops" => { TARGET: 2, SRC: "simt_systim2",
 				DEFS: "-DHRT_CONFIG1 -DHOOK_HRT_EVENT -DUSE_64BIT_OPS" },
-  "systim3_64ops" => { TARGET: 1, SRC: "simt_systim3",
+  "systim3_64ops" => { TARGET: 2, SRC: "simt_systim3",
 				DEFS: "-DHRT_CONFIG1 -DHOOK_HRT_EVENT -DUSE_64BIT_OPS" },
 
   # 性能評価プログラム
-  "perf0" => { CDL: "perf_pf" },
-  "perf1" => { CDL: "perf_pf" },
-  "perf2" => { CDL: "perf_pf" },
-  "perf3" => { CDL: "perf_pf" },
-  "perf4" => { CDL: "perf_pf" },
-  "perf5" => { CDL: "perf_pf" },
+  "perf0" => { TARGET: 1, CDL: "perf_pf", NK_DEFS: "-DHIST_INVALIDATE_CACHE" },
+  "perf1" => { TARGET: 1, CDL: "perf_pf", NK_DEFS: "-DHIST_INVALIDATE_CACHE" },
+  "perf2" => { TARGET: 1, CDL: "perf_pf", NK_DEFS: "-DHIST_INVALIDATE_CACHE" },
+  "perf3" => { TARGET: 1, CDL: "perf_pf", NK_DEFS: "-DHIST_INVALIDATE_CACHE" },
+  "perf4" => { TARGET: 1, CDL: "perf_pf", NK_DEFS: "-DHIST_INVALIDATE_CACHE" },
+  "perf5" => { TARGET: 1, CDL: "perf_pf", NK_DEFS: "-DHIST_INVALIDATE_CACHE" },
 
   # ARM向けテストプログラム
-  "arm_cpuexc" => { SRC: "arm_cpuexc", SRCDIR: "arch/arm_gcc/test" },
-  "arm_fpu1" => { TARGET: 2, SRC: "arm_fpu1", SRCDIR: "arch/arm_gcc/test" },
+  "arm_cpuexc1" => { SRC: "arm_cpuexc1", SRCDIR: "arch/arm_gcc/test" },
+  "arm_fpu1" => { TARGET: 3, SRC: "arm_fpu1", SRCDIR: "arch/arm_gcc/test" },
 }
 
 #
 #  カーネルライブラリの作成
 #
-def BuildKernel
-  if !Dir.exist?("KERNELLIB")
-    Dir.mkdir("KERNELLIB")
+def BuildKernel(target, mkdirFlag=false)
+  return unless $targetOptions.has_key?(target)
+
+  kernelDir = "KERNELLIB" + target.to_s
+  if !Dir.exist?(kernelDir)
+    if mkdirFlag
+      Dir.mkdir(kernelDir)
+    else
+      return
+    end
   end
 
-  Dir.chdir("KERNELLIB") do
-    puts("== building: KERNELLIB ==")
+  Dir.chdir(kernelDir) do
+    puts("== building: #{kernelDir} ==")
     configCommand = "ruby #{$usedSrcDir}/configure.rb"
-    configCommand += " #{$targetOptions[0]}"
     configCommand += " -f"
+    configCommand += " #{$targetOptions[target]}"
     puts(configCommand)
     system(configCommand)
     system("make libkernel.a")
     if File.exist?("Makefile.bak")
       File.delete("Makefile.bak")
     end
+  end
+end
+
+#
+#  全カーネルライブラリの作成
+#
+def BuildAllKernel()
+  $targetOptions.keys.each do |target|
+    BuildKernel(target)
   end
 end
 
@@ -195,8 +253,15 @@ def BuildTest(test, testSpec, mkdirFlag=false)
       configCommand += " -a #{$usedSrcDir}/test"
     end
 
-    if (!testSpec.has_key?(:TARGET) || testSpec[:TARGET] == 0)
-      configCommand += " -L ../KERNELLIB"
+    if !testSpec.has_key?(:DEFS)
+      if (testSpec.has_key?(:TARGET))
+        kernelDir = "KERNELLIB" + testSpec[:TARGET].to_s
+      else
+        kernelDir = "KERNELLIB0"
+      end
+      if Dir.exist?("../" + kernelDir)
+        configCommand += " -L ../" + kernelDir
+      end
     end
     if testSpec.has_key?(:SRC)
       configCommand += " -A #{testSpec[:SRC]}"
@@ -211,9 +276,9 @@ def BuildTest(test, testSpec, mkdirFlag=false)
     else
       configCommand += " -C test_pf.cdl"
     end
-    if testSpec.has_key?(:SYSLIB)
+    if testSpec.has_key?(:SYSOBJ)
       configCommand += " -S \"" \
-			+ testSpec[:SYSLIB].split(/\s+/).map{|f| f+".o"}.join(" ") \
+			+ testSpec[:SYSOBJ].split(/\s+/).map{|f| f+".o"}.join(" ") \
 			+ "\""
     end
     if testSpec.has_key?(:APPLOBJ)
@@ -227,6 +292,9 @@ def BuildTest(test, testSpec, mkdirFlag=false)
     if testSpec.has_key?(:DEFS)
       configCommand += " -O \"#{testSpec[:DEFS]}\""
     end
+    if testSpec.has_key?(:NK_DEFS)
+      configCommand += " -O \"#{testSpec[:NK_DEFS]}\""
+    end
     puts(configCommand)
     system(configCommand)
     system("make")
@@ -239,7 +307,7 @@ end
 #
 #  全テストプログラムの作成
 #
-def BuildAll
+def BuildAllTest()
   TEST_SPEC.each do |test, testSpec|
     BuildTest(test, testSpec)
   end
@@ -267,7 +335,7 @@ end
 #
 #  全テストプログラムの実行
 #
-def ExecAll
+def ExecAllTest()
   TEST_SPEC.each do |test, testSpec|
     ExecTest(test, testSpec)
   end
@@ -276,11 +344,23 @@ end
 #
 #  カーネルライブラリのクリーン
 #
-def CleanKernel
-  if Dir.exist?("KERNELLIB")
-    Dir.chdir("KERNELLIB") do
+def CleanKernel(target)
+  return unless $targetOptions.has_key?(target)
+
+  kernelDir = "KERNELLIB" + target.to_s
+  if Dir.exist?(kernelDir)
+    Dir.chdir(kernelDir) do
       system("make clean")
     end
+  end
+end
+
+#
+#  全カーネルライブラリのクリーン
+#
+def CleanAllKernel()
+  $targetOptions.keys.each do |target|
+    CleanKernel(target)
   end
 end
 
@@ -301,7 +381,7 @@ end
 #
 #  全テストプログラムのクリーン
 #
-def CleanAll
+def CleanAllTest()
   TEST_SPEC.each do |test, testSpec|
     CleanTest(test, testSpec)
   end
@@ -329,41 +409,79 @@ $targetOptions = {}
 File.open("TARGET_OPTIONS") do |file|
   file.each_line.with_index do |line, index|
     line.chomp!
-    $targetOptions[index] = line
+    if line != ""
+      $targetOptions[index] = line
+    end
   end
 end
 
 #
 #  パラメータで指定された処理の実行
 #
-if ARGV.size == 0
-  BuildKernel()
-  BuildAll()
-  ExecAll()
-else
-  ARGV.each do |param|
-    case param
-    when "kernel"
-      BuildKernel()
-    when "build"
-      BuildAll()
-    when "exec"
-      ExecAll()
-    when "all"
-      BuildAll()
-      ExecAll()
-    when "clean_kernel"
-      CleanKernel()
-    when "clean"
-      CleanKernel()
-      CleanAll()
+$build_only = false
+$exec_only = false
+$clean_flag = false
+$proc_flag = false
+
+ARGV.each do |param|
+  case param
+  when "build"
+    $build_only = true
+  when "exec"
+    $exec_only = true
+  when "clean"
+    $clean_flag = true
+
+  when "kernel"
+    if ($clean_flag)
+      CleanAllKernel()
     else
-      if TEST_SPEC.has_key?(param)
-        BuildTest(param, TEST_SPEC[param], true)
-        ExecTest(param, TEST_SPEC[param])
-      else
-        puts("invalid parameter: #{param}")
-      end
+      BuildAllKernel() unless $exec_only
+      # カーネルには，execはない
     end
+    $proc_flag = true
+
+  when /^kernel([0-9]+)$/
+    target = $1.to_i
+    if ($clean_flag)
+      CleanKernel(target)
+    else
+      BuildKernel(target, true) unless $exec_only
+      # カーネルには，execはない
+    end
+    $proc_flag = true
+
+  when "all"
+    if ($clean_flag)
+      CleanAllTest()
+    else
+      BuildAllTest() unless $exec_only
+      ExecAllTest() unless $build_only
+    end
+    $proc_flag = true
+
+  else
+    if TEST_SPEC.has_key?(param)
+      if ($clean_flag)
+        CleanTest(param, TEST_SPEC[param])
+      else
+        BuildTest(param, TEST_SPEC[param], true) unless $exec_only
+        ExecTest(param, TEST_SPEC[param]) unless $build_only
+      end
+    else
+      puts("invalid parameter: #{param}")
+    end
+    $proc_flag = true
+  end
+end
+if !$proc_flag
+  # デフォルトの処理対象（kernelとall）
+  if ($clean_flag)
+    CleanAllKernel()
+    CleanAllTest()
+  else
+    BuildAllKernel() unless $exec_only
+    BuildAllTest() unless $exec_only
+    ExecAllTest() unless $build_only
   end
 end
